libopie2: install include.pro into palmtopdir
Installing files into the sysroot during do_install is not a good idea -
in particular this breaks with OE-core's machine-specific sysroots.
Instead we now install it to ${D}${palmtopdir} and package it with the
other development files, and apply a sed hack to all .pro files within
opie.bbclass to ensure it gets found when building each package.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>