qfish2: fix broken SRC_URI and non-ia32 compilation
Use a more up-to-date official source; this is actually a different
archive and hence the checksums changed - archive contents are mostly
the same except for the removal of some erroneously shipped files.

A compilation fix was also needed - a tools/bin2c 32-bit x86 executable
was being shipped and used during the build, which does not work on
other architectures (e.g. x86-64). Deleting the file ensures the
executable itself is built during the build process.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>