classes/opie: fix ownership of files installed in do_opie_install
We're adding our own install task here, this needs to be marked as
fakeroot so that it runs under pseudo and thus files are installed as
root. In the process I discovered that a function exported with
EXPORT_FUNCTIONS marked as fakeroot doesn't actually end up as a
fakeroot task - you have to force it using d.setVarFlag() from anonymous
python; this is probably a bug in BitBake.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>