opie-systemtime: drop unnecessary FILES line
This FILES line isn't actually doing anything useful - we're not
installing the file to /etc, we're installing it to ${palmtopdir}/etc.
I'm not immediately sure if that's correct, but even if it wasn't we
wouldn't need to add ${sysconfdir} anyway, and in any case it causes a
warning, so just drop it for now.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>